0-print_list.c file

+++ 0x12-singly_linked_lists/0-print_list.c
@@ -9,14 +9,14 @@
  */
 size_t print_list(const list_t *h)
 {
-	size_t node_count = 0;
+	unsigned int node_count = 0;
 
 	while (h)
 	{
 		if (!h->str)
 			printf("[0] (nil)\n");
 		else
-			printf("[%u] %s\n", h->len, h->str);
+			printf("[%d] %s\n", h->len, h->str);
 		h = h->next;
 		node_count++;
 	}
